WorkFile and XWFile classes deleted; refactoring
